+++ ScramcastSubServer.h
@@ -14,7 +14,7 @@ class ScramcastSubServer: public ScramcastServer {
 public:
 	ScramcastSubServer(int tcp_sock, int scram_sock, int hostId);
 	virtual ~ScramcastSubServer();
-	virtual int32_t AddMemoryWatch(u_int8_t NetId, u_int32_t Offset, u_int32_t Length, u_int32_t resolution);
+	virtual int32_t AddMemoryWatch(u_int32_t NetId, u_int32_t Offset, u_int32_t Length, u_int32_t resolution);
 private:
 	int _clntSock;
 };
